// ts/social-react-core/gen/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Duration: { input: string; output: string; }
  InputTimestamp: { input: any; output: any; }
  Timestamp: { input: string; output: string; }
};

export type ApiAddressInput = {
  address?: InputMaybe<Scalars['String']['input']>;
  city?: InputMaybe<Scalars['String']['input']>;
  country?: InputMaybe<Scalars['String']['input']>;
  zip?: InputMaybe<Scalars['String']['input']>;
};

export type ApiCursorInput = {
  after?: InputMaybe<Scalars['String']['input']>;
  before?: InputMaybe<Scalars['String']['input']>;
  first?: InputMaybe<Scalars['Int']['input']>;
  last?: InputMaybe<Scalars['Int']['input']>;
};
// ...
export enum ApiEntityState {
  EntityStateBlocked = 'ENTITY_STATE_BLOCKED',
  EntityStateDeleted = 'ENTITY_STATE_DELETED',
  EntityStateUnspecified = 'ENTITY_STATE_UNSPECIFIED'
}
// ...
export type AuthAccount = {
  __typename?: 'AuthAccount';
  acceptedTerms: Array<AuthTermsVersion>;
  birthday?: Maybe<AuthDate>;
  channelCreationEligibility: ChannelChannelCreationEligibility;
  createdAt: Scalars['Timestamp']['output'];
  email: Scalars['String']['output'];
  emailVerifiedAt?: Maybe<Scalars['Timestamp']['output']>;
  externalIds: Array<AuthIdentity>;
  flags: Array<AuthAccountStatusFlag>;
  isBot: Scalars['Boolean']['output'];
  labels: Array<Scalars['String']['output']>;
  marketingConsent: AuthConsentStatus;
  matureRatedContentAllowed: Scalars['Boolean']['output'];
  pendingAgreements: Array<AgreementAgreementRevision>;
  roles: Array<AuthPlatformRole>;
  signupLocation: Scalars['String']['output'];
  signupOrigin?: Maybe<AuthSignupOrigin>;
  staff: Scalars['Boolean']['output'];
  state: ApiEntityState;
  temporary: Scalars['Boolean']['output'];
  uid: Scalars['String']['output'];
};
// ...
export type BadgeBadge = {
  __typename?: 'BadgeBadge';
  level: Scalars['Int']['output'];
  nextLevelAt: Scalars['Timestamp']['output'];
  type: BadgeBadgeType;
};
// ...
export enum ChannelAppealStatus {
  AppealStatusAccepted = 'APPEAL_STATUS_ACCEPTED',
  AppealStatusDeclined = 'APPEAL_STATUS_DECLINED',
  AppealStatusPending = 'APPEAL_STATUS_PENDING',
  AppealStatusUnspecified = 'APPEAL_STATUS_UNSPECIFIED'
}
// ...
export type ChannelBanAppealInfo = {
  __typename?: 'ChannelBanAppealInfo';
  appealText: Scalars['String']['output'];
  channelId: Scalars['ID']['output'];
  createdAt: Scalars['Timestamp']['output'];
  reviewer: ProfileProfile;
  reviewerComment: Scalars['String']['output'];
  reviewerId: Scalars['ID']['output'];
  status: ChannelAppealStatus;
  userId: Scalars['ID']['output'];
};
// ...
export type ChannelChannel = {
  __typename?: 'ChannelChannel';
  activeStream?: Maybe<ChannelStreamSummary>;
  channelFriends: FriendsChannelFriends;
  createdAt: Scalars['Timestamp']['output'];
  currentChatId?: Maybe<Scalars['String']['output']>;
  currentStream?: Maybe<ChannelStream>;
  currentStreamId: Scalars['ID']['output'];
  description: Scalars['String']['output'];
  features: ChannelChannelFeatures;
  followerCount: Scalars['Int']['output'];
  following: Scalars['Boolean']['output'];
  game: GameGame;
  gameId: Scalars['ID']['output'];
  id: Scalars['ID']['output'];
  isPublic: Scalars['Boolean']['output'];
  itemStats: Array<ItemItemStat>;
  links: Array<ChannelChannelLink>;
  liveStatus: ChannelLiveStatus;
  logo: Scalars['String']['output'];
  matchOngoing: Scalars['Boolean']['output'];
  matureRatedContent: Scalars['Boolean']['output'];
  monetizationSettings: ChannelMonetizationSettings;
  name: Scalars['String']['output'];
  offlineBanner: Scalars['String']['output'];
  offlineBannerUrl: Scalars['String']['output'];
  playedGameIds: Array<Scalars['String']['output']>;
  priority: Scalars['Int']['output'];
  reportingId: Scalars['ID']['output'];
  riskTier: ChannelRiskTier;
  state: ApiEntityState;
  streamedGames?: Maybe<Array<GameGame>>;
  streamer: ProfileProfile;
  streamerId: Scalars['ID']['output'];
  subscriberCount: Scalars['Int']['output'];
  subscription?: Maybe<SubscriptionChannelSubscription>;
  subscriptionConfig?: Maybe<SubscriptionChannelSubscriptionConfig>;
  suspension: ChannelSuspension;
  thumbnail: Scalars['String']['output'];
  title: Scalars['String']['output'];
  userBanStatus: ChannelUserBanStatus;
  viewerCount: Scalars['Int']['output'];
};
// ...
export type ChannelChannelFeatures = {
  __typename?: 'ChannelChannelFeatures';
  challenges: ChannelChallengesFeatureStatus;
  noicePredictions: ChannelNoicePredictionsFeatureStatus;
  streaming: ChannelStreamingFeatureStatus;
};

export type ChannelChannelLink = {
  __typename?: 'ChannelChannelLink';
  name: Scalars['String']['output'];
  type: ChannelChannelLinkLinkType;
  url: Scalars['String']['output'];
};
// ...
export enum ChannelLiveStatus {
  LiveStatusLive = 'LIVE_STATUS_LIVE',
  LiveStatusOffline = 'LIVE_STATUS_OFFLINE',
  LiveStatusUnlisted = 'LIVE_STATUS_UNLISTED',
  LiveStatusUnspecified = 'LIVE_STATUS_UNSPECIFIED'
}
// ...
export type ChannelMonetizationSettings = {
  __typename?: 'ChannelMonetizationSettings';
  channelId: Scalars['ID']['output'];
  eligible: Scalars['Boolean']['output'];
  enabled: Scalars['Boolean']['output'];
};
// ...
export enum ChannelRiskTier {
  RiskTier_0 = 'RISK_TIER_0',
  RiskTier_1 = 'RISK_TIER_1',
  RiskTier_2 = 'RISK_TIER_2',
  RiskTierUnspecified = 'RISK_TIER_UNSPECIFIED'
}

export type ChannelStream = {
  __typename?: 'ChannelStream';
  challengesEnabled: Scalars['Boolean']['output'];
  channelId: Scalars['ID']['output'];
  matureRatedContent: Scalars['Boolean']['output'];
  noicePredictionsEnabled: Scalars['Boolean']['output'];
  segments: Array<ChannelStreamSegment>;
  serverRenderingEnabled: Scalars['Boolean']['output'];
  streamId: Scalars['ID']['output'];
  transcodingEnabled: Scalars['Boolean']['output'];
};
// ...
export type ChannelStreamSummary = {
  __typename?: 'ChannelStreamSummary';
  challengesEnabled: Scalars['Boolean']['output'];
  channelId: Scalars['ID']['output'];
  endTime?: Maybe<Scalars['Timestamp']['output']>;
  gameId: Scalars['ID']['output'];
  matureRatedContent: Scalars['Boolean']['output'];
  noicePredictionsEnabled: Scalars['Boolean']['output'];
  serverRenderingEnabled: Scalars['Boolean']['output'];
  startTime: Scalars['Timestamp']['output'];
  streamId: Scalars['ID']['output'];
  transcodingEnabled: Scalars['Boolean']['output'];
};
// ...
export type ChannelSuspension = {
  __typename?: 'ChannelSuspension';
  description: Scalars['String']['output'];
  moderator?: Maybe<ProfileProfile>;
  reason: ChannelSuspensionReason;
  suspendedAt: Scalars['Timestamp']['output'];
  suspendedBy: Scalars['String']['output'];
  until?: Maybe<Scalars['Timestamp']['output']>;
};
// ...
export type ChannelUserBanStatus = {
  __typename?: 'ChannelUserBanStatus';
  appeal?: Maybe<ChannelBanAppealInfo>;
  banned: Scalars['Boolean']['output'];
  bannedAt?: Maybe<Scalars['Timestamp']['output']>;
  channel: ChannelChannel;
  channelId: Scalars['ID']['output'];
  description: Scalars['String']['output'];
  moderator?: Maybe<ProfileProfile>;
  moderatorId: Scalars['ID']['output'];
  user: ProfileProfile;
  userId: Scalars['ID']['output'];
  violation: ChannelViolation;
};
// ...
export enum ChannelViolation {
  ViolationOther = 'VIOLATION_OTHER',
  ViolationSpam = 'VIOLATION_SPAM',
  ViolationUnspecified = 'VIOLATION_UNSPECIFIED'
}
// ...
export type FriendsChannelFriends = {
  __typename?: 'FriendsChannelFriends';
  channelId: Scalars['ID']['output'];
  totalCount: Scalars['Int']['output'];
  users: Array<FriendsUser>;
};
// ...
export type GameGame = {
  __typename?: 'GameGame';
  activeSeason: GameSeason;
  activeSeasonId: Scalars['ID']['output'];
  backdropUrl: Scalars['String']['output'];
  challengesEnabled: Scalars['Boolean']['output'];
  coverImageUrl: Scalars['String']['output'];
  gameEventsSource: GameGameEventsSource;
  iconUrl: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  name: Scalars['String']['output'];
  noicePredictionsEnabled: Scalars['Boolean']['output'];
  otherNames: Array<Scalars['String']['output']>;
  progression: GameUserProgression;
  publicAccess: Scalars['Boolean']['output'];
};
// ...
export type GoogleProtobufEmpty = {
  __typename?: 'GoogleProtobufEmpty';
  emptyTypeWorkaround: Scalars['Boolean']['output'];
};
// ...
export type ItemItemStat = {
  __typename?: 'ItemItemStat';
  counts: Array<ItemItemCount>;
  gameId: Scalars['ID']['output'];
};
// ...
export type ProfileProfile = {
  __typename?: 'ProfileProfile';
  account?: Maybe<AuthAccount>;
  avatarConfig?: Maybe<ProfileProfileAvatarConfig>;
  /**
   * contains a concatenated list of different avatar urls (use avatars object instead)
   * @deprecated field is deprecated
   */
  avatarUrl: Scalars['String']['output'];
  avatars?: Maybe<ProfileProfileAvatars>;
  badges: Array<BadgeBadge>;
  bio: Scalars['String']['output'];
  canChangeUsernameAt?: Maybe<Scalars['Timestamp']['output']>;
  channel?: Maybe<ChannelChannel>;
  discordUsername?: Maybe<Scalars['String']['output']>;
  /** @deprecated field is deprecated */
  displayName: Scalars['String']['output'];
  friends: Array<ProfileProfile>;
  friendshipStatus: FriendsFriendshipStatus;
  isNewUsername: Scalars['Boolean']['output'];
  lastSeen?: Maybe<Scalars['Timestamp']['output']>;
  onlineStatus: ProfilePresenceStatus;
  playedGames: Array<ProfilePlayedGame>;
  preferredColor: ProfileColor;
  settings?: Maybe<ProfileProfileSettings>;
  state: ApiEntityState;
  stats: PlayerStatsPlayerStats;
  temporary: Scalars['Boolean']['output'];
  userId: Scalars['ID']['output'];
  userTag: Scalars['String']['output'];
  usernameHistory?: Maybe<Array<ProfileUsernameChange>>;
  visibility: ProfileProfileVisibility;
};
// ...
export type ProfileProfileAvatarConfig = {
  __typename?: 'ProfileProfileAvatarConfig';
  model: AvatarAvatar;
  modelId: Scalars['ID']['output'];
};

export type ProfileProfileAvatars = {
  __typename?: 'ProfileProfileAvatars';
  avatar2D: Scalars['String']['output'];
  avatar3D: Scalars['String']['output'];
  avatarFullbody: Scalars['String']['output'];
  avatarGender: Scalars['String']['output'];
};
// ...
export type QueryChannelBanUserStatusArgs = {
  channelId?: InputMaybe<Scalars['ID']['input']>;
  userId?: InputMaybe<Scalars['ID']['input']>;
};


export type QueryChannelBanUserStatusesArgs = {
  channelIds?: InputMaybe<Array<Scalars['String']['input']>>;
  userId?: InputMaybe<Scalars['ID']['input']>;
};


export type QueryChannelBannedUsersArgs = {
  channelId?: InputMaybe<Scalars['ID']['input']>;
  cursor?: InputMaybe<ApiCursorInput>;
};
// ...
export type SubscriptionChannelSubscription = {
  __typename?: 'SubscriptionChannelSubscription';
  activatedAt?: Maybe<Scalars['Timestamp']['output']>;
  cancelReason: SubscriptionChannelSubscriptionCancelReason;
  cancelledAt?: Maybe<Scalars['Timestamp']['output']>;
  channel: ChannelChannel;
  channelId: Scalars['ID']['output'];
  createdAt: Scalars['Timestamp']['output'];
  expiresAt?: Maybe<Scalars['Timestamp']['output']>;
  externalReference: Scalars['String']['output'];
  giverId: Scalars['ID']['output'];
  id: Scalars['ID']['output'];
  paymentFailedAt?: Maybe<Scalars['Timestamp']['output']>;
  paymentSucceededAt?: Maybe<Scalars['Timestamp']['output']>;
  provider: SubscriptionChannelSubscriptionProvider;
  renewedAt?: Maybe<Scalars['Timestamp']['output']>;
  state: SubscriptionChannelSubscriptionState;
  terminatedAt?: Maybe<Scalars['Timestamp']['output']>;
  tier: Scalars['Int']['output'];
  user: ProfileProfile;
  userId: Scalars['ID']['output'];
};
// ...
export type SubscriptionChannelSubscriptionConfig = {
  __typename?: 'SubscriptionChannelSubscriptionConfig';
  channelId: Scalars['ID']['output'];
  subscriptionsEnabled: Scalars['Boolean']['output'];
  tiers: Array<SubscriptionChannelSubscriptionTier>;
};
// ...
export type SocialChannelUnbanMutationVariables = Exact<{
  userId: Scalars['ID']['input'];
  channelId: Scalars['ID']['input'];
}>;


export type SocialChannelUnbanMutation = { __typename?: 'Mutation', unbanChannelUser?: { __typename?: 'GoogleProtobufEmpty', emptyTypeWorkaround: boolean } | null };